\name{formatdfExport}
\alias{formatdfExport}
\title{
    Format a data frame or a matrix for export
}
\description{
    Rcmdr interface to format matrices and data frames for 
    export to LaTeX or HTML code.
}
\usage{
    formatdfExport()
}
\details{
The graphical interface to \code{\link[Hmisc]{format.df}} allows to 
format any matrix or data.frame from the current workspace.  The printed 
object can subsequently be retrieved for exporting to 
\code{LaTeX} or \code{HTML}. 

}
\value{
    \code{NULL}
}
\author{Liviu Andronic \email{landronimirc@gmail.com}}
%\references{
%}
\seealso{\code{\link[Hmisc]{format.df}}}
\keyword{ print }
