\name{Utilities}
\alias{Utilities}
\alias{listMatrixObjects}
\title{
    Rcmdr.Export utility functions
}
\description{
    Rcmdr.Export utility functions
}
\usage{
    listMatrixObjects(envir=.GlobalEnv, ...)
}
\arguments{
% adapted from Rcmdr utilities.Rd
    \item{envir}{the environment to be searched; should generally be left at the default.}
    \item{\dots}{currently ignored.}

}
\details{
This section lists the functions used to support the graphical dialogues.  

}
\value{
    \code{NULL}
}
\keyword{misc}
